Extract chain methods
